Patch from Daniel, with major cleanup by me: Added a common base class on the heap allocated objects (to replace boost::noncopyable). This class enforces the non-copyability as well as uses a conservative heuristic to keep track of when the object has been deleted but is still in use and to try to throw an exception when such has been detected.
We don't currently have issues with object ownership in IMP, but they are a
pain to track down when they occur as Python generally just unceremoniously
bombs. And it gets rid of the funny IMP/boost files.

A few other bounds checks and null pointer checks have been added also.


git-svn-id: https://svn.salilab.org/imp/trunk@281 e91f8c7b-0e32-4433-a9ac-4502d39a4eab