Patch from Daniel: I cleaned up the bond decorator code. Among other things I removed the length parameter with the expectation that I will add another similar set of decorators for custom connectors (not quite sure what to call them). These will represent arbitrary length attachments between particles. I also added some test code for the bond decorators.
The connectors will have a length and stiffness parameter and will be handled
all together by a restraint (along the lines of what was discussed with the
non-bonded list). Normal inter-atomic bonds could be handled by generating
the appropriate lengths from the names of the atoms involved in each atomic
bond and adding the connector information.


git-svn-id: https://svn.salilab.org/imp/trunk@320 e91f8c7b-0e32-4433-a9ac-4502d39a4eab