Patch from Daniel, with major cleanup by me: Implement Monte Carlo. It is structured as an optimizer coupled with a set of Movers which propose moves. The optimizer holds a collection of Movers and a local optimizer (which can be used to mix local optimization with Monte Carlo). Current movers take a set of particles and a set of attributes and perturb all of those attributes on each of the particles either uniformly in a a ball or with a normal distribution.
A class called MoveGuard keeps track of the information necessary to  
revert if the monte carlo step is rejected. Note that as it stands,  
writing a Mover in Python would take noticeably more work as the macro  
does a good bit of work.

The current implementation uses the boost random number library, thus we now
depend on boost to build IMP. Unfortunately, the current boost version
generates a warning in one of its headers and so -Werror has to be disabled.


git-svn-id: https://svn.salilab.org/imp/trunk@363 e91f8c7b-0e32-4433-a9ac-4502d39a4eab