Patch from Daniel: removed the unused stats code in ModelData.h. We can do the same things through ScoreStates (see MaxChangeScoreState)
It also removes the flag which is set when particles are activated or
deactivated since that is not used any more. If we have one, it should be in
Model anyway.

And moves the vector include to after the IMP includes so it doesn't mask
any missing vector includes in them.


git-svn-id: https://svn.salilab.org/imp/trunk@379 e91f8c7b-0e32-4433-a9ac-4502d39a4eab