Patch from Daniel, with cleanup by me: Remove ModelData. The interesting changes are that Optimizer now has a couple of ModelData-esque helper methods to enable the ConjugateGradients and SteepestDescent optimizers to work with minimal changes (these allow easy access to the optimized attributes). Modifying the optimizers could make them more efficient (by grouping the attributes by particle), but that can wait.
Modified and cleaned up MolecularDynamics a bit. The main change is it now
skips particles which don't have optimizable x,y,z. This seems more consistent
with the idea of an optimizer (it doesn' have to actually improve every
optimizable attribute) and we need the optimizer to work when there are other
sorts of particles floating around. It uses an IMP_LIST in order to simplify
moving to reference counted particles if we do that.


git-svn-id: https://svn.salilab.org/imp/trunk@435 e91f8c7b-0e32-4433-a9ac-4502d39a4eab