Replace the code from the Monte Carlo NormalMover::generate_move method that uses boost::uniform_01 with code that uses boost::variate_generator instead.
This is because uniform_01 takes the random number generator by value rather
than reference; thus, it creates a temporary copy of it and doesn't update the
generator state (so that every time you call the method, you get the same
result, unless you call some other method somewhere which _does_ update
the RNG).

We now pass by reference so that the RNG state is updated, as was done for the
MD code in r454.


git-svn-id: https://svn.salilab.org/imp/trunk@455 e91f8c7b-0e32-4433-a9ac-4502d39a4eab