Patch from Daniel, with minor cleanup by me: Adds the ability to remove particle attributes. It also adds more thorough checking of NaNs in the float attributes. In addition, the rearrangement creates a natural iterator through the optimized float keys which the Optimizer now uses.
Internally, it changes how attributes are stored, replacing the valid bit
with a sentinel value. This works nicely for floats and derivs where we can
use NaN and for particles where we can use NULL. It also reduces the memory
usage for those by a bit since the bools couldn't be packed well before.

For ints, it is a little less nice as it currently uses
numeric_limits<int>::max as a sentinel. For strings it is a particular weird
string. For both of those we can easily improve things if it becomes a
problem (either due to people wanting to use int max as an attribute value
or the memory allocation associated with the strings becoming significant).


git-svn-id: https://svn.salilab.org/imp/trunk@577 e91f8c7b-0e32-4433-a9ac-4502d39a4eab