Patch from Daniel: As part of the discussed reorg of vector and matrix stuff, switch to a d-dimensional implementation for vectors.
In order to do nice error checking it adds a preprocessor macro to the _wrap
compilation which is used to disable the compile time checks (since SWIG
instantiates all methods, for obvious reasons) and replace them with runtime
checks.

For some reason I need a typedef for Vector3D in each of the module .i files.
I don't see why, but I can't get things to work without it.


git-svn-id: https://svn.salilab.org/imp/trunk@869 e91f8c7b-0e32-4433-a9ac-4502d39a4eab