Changes to support hierarchy of particles.
1. Change DiscreteSampler::move2state not to be const, effects:
PermutationSampler.h,CartesianProductSampler.h,JNode.h
SimpleDiscreteSampler.h,DiscreteSampler.h and the .cpp

2. In RestraintGraph change the interface of initialize_potentials.
The idea is that RestraintGraph now passes the particles in the hierarchy 
level encoded in the nodes.

3. In JNode.h change the interface of realize

4. DominoOptimizer:
change the mechanism in which the restraints are realized. 
Instead of taking the restraints from the Model, 
the relevant restraints should be provided. The reason is that we only 
need the particles at the level encoded in the junction tree. 
I wanted to avoid any information of hierarhcy in the domino optimizer. 
The only place where this information is considered is in the 
discrete samplers, for example for the move2state function.
The interface changed, for example the function add_restraint was added.

git-svn-id: https://svn.salilab.org/imp/trunk@1308 e91f8c7b-0e32-4433-a9ac-4502d39a4eab