Upgrade the doxygen.conf-in
Also, fix the SConscript so that it handles the bits to be replaced being on the same line as the variable name in teh doxygen.conf template. That way we don't have to patch the template each time we upgrade. Also (yeah, too many disparate changes, should use git's interactive commit) turn on source browsing so the place where functions are defined is shown. This is more info than we want, but does include the header to include. And, as a trial, turn the tree view back on.

git-svn-id: https://svn.salilab.org/imp/trunk@15025 e91f8c7b-0e32-4433-a9ac-4502d39a4eab