simplify the dependency tracking code
We now write a file out per external dependency in build/dependencies. This file has all the info that we need for the dependency. Later, we can add the files to SCons as dependencies for things like config headers. It also creates hooks, not currently implemented to provide code to build external dependencies within IMP

git-svn-id: https://svn.salilab.org/imp/trunk@15435 e91f8c7b-0e32-4433-a9ac-4502d39a4eab