Don't force Restraints and ScoreStates to divide inputs
No longer for Restraints and ScoreStates to divide their inputs into Particles and Containers as this provides no way of being dependent directly on ScoreStates. It also complicates a great deal of code. To maintain backwards compatibility, we fork the helper macros and provide specialized base classes for python.

Note, this for some reason breaks the ability to use get_from() with IMP.Restraint within python. As doing such is pretty useless, it doesn't seem worth investigating.

git-svn-id: https://svn.salilab.org/imp/trunk@15585 e91f8c7b-0e32-4433-a9ac-4502d39a4eab