Update evaluation architecture to support OpenMP evaluation
This patch updates how restraints are evaluated in order to facilitate OpenMP support. The main restraint evaluation mechanism is via the ScoreAccumulator which takes a piece of the score an adds it to the total which is stored elsewhere in a manner similar to DerivativeAccumulators. The change is backwards compatibility except for ScoringFunctions, but we don't believe there to be any outside of the IMP core.

git-svn-id: https://svn.salilab.org/imp/trunk@15791 e91f8c7b-0e32-4433-a9ac-4502d39a4eab