Try to get line numbers correct; closes #388.
doxygen warnings or errors refer to line numbers in the filtered text,
not the original file. Try to pad the filtered output so that function
and class definitions in the filtered text end up as close to their positions
in the original file as possible. (We can't do a perfect job because the file
has to be reordered - Python docstrings come *after* the class or function
definition, while doxygen expects to see the comments *before* the def.)