Add an extra parameter to the IMP_GRAPH macros to control how vertices are shown
I don't expect that these are used outside out code base as this point and the update is somple, so I think the compatibility break is ok. This parameter makes it so that we can, eg, show the type in the Dependency graph, which makes that graph much easier to read.
git status
close #397, close #396, close #395

minor changes to be rebased

minor changes to be rebased

minor changes to be rebased