remove %template for Vector3Ds
it isn't needed for correctness and breaks compilation on Mac OS 10.9. The breakage may be swig's fault (it seems to be trying to instantiate vector::erase(value_type)). However, it may make the function that uses it create_sphere_dots (which should be get_sphere_dots) a bit slower when called from python due to converting the data to a python list.