Add multiresolution and states to IMP.atom
Add the IMP.atom.Resolution decorator which implements multiresolution
as parallel hierarchies. And add support to selection also. This
removes the radius criteria from Selection, but no one seemed to be
able to figure that one out. We can add a backwards compat method if
needed for that, but not having it is way simpler.

Also, add IMP.atom.State decorator.

Link both with RMF.

close #647
close #648