Use IMP info if we don't have useful RMF info.
When linking to an RMF file, if we haven't read a frame yet
we can only see the static values, and for nonrigid members
that are also rigid bodies, there are no such values for
translation/rotation, so they are not read in correctly.
But IMP knows that the member is nonrigid and is a rigid
body, so use that information.

Closes #845.