Don't require Python extensions to build docs in static build.
We normally require the Python extensions to be built to build the
documentation (since Python-only classes are also included in the
docs) but this won't work with the static build (since Python
extensions don't work there). Rather than breaking the docs entirely
in this case, just document the C++ side of things.