Treat nonmember rigid body children as nonrigid.
In the case where a particle is both a Hierarchy and a RigidBody,
other particles could be children of the Hierarchy but not members
of the RigidBody. This possibility is not covered in the current
code and results in an exception.

Work around this by treating particles that are children but not
members as if they are nonrigid members. This results in tidier RMF
files than using the external rigid body mechanism (which duplicates
the rigid body's reference frame for every particle in the body).
The (minor) drawback is that when the RMF is read back in, these
particles will become "real" nonrigid members.

Closes #879.