Don't pass CGAL_CXX_FLAGS to CGAL setup.
We use CGAL_CXX_FLAGS to control compilation of the IMP.cgal
module, but CGAL uses it too. This means that the second time
we run cmake CGAL_CXX_FLAGS gets incorporated into CMAKE_CXX_FLAGS.
This is turn forces everything to be rebuilt (since the command line
has changed). This is most obviously a problem in the .rpm and .deb
builds, where cmake is rerun to add Python 3 support; we only want
to rebuild the Python wrappers, not all of IMP.