Ignore some 'duplicate anchor' warnings.
These seem to be referring to anchors in the kernel's README.
It's not clear why they are being duplicated from the kernel
tagfile into those for other tagfiles (possibly a doxygen bug?)
but the output HTML appears correct. Suppress these particular
warnings so that it's easier to see genuine problems.