Return vector to Python rather than taking a reference.
Taking a reference is an unusual way of handling lists in Python,
and to make it work here required exposing std::vector<float/real>
to Python, which was ugly (the types did not behave like real
Python lists). Return new vectors instead, which are automatically
handled by our existing wrappers and converted into Python lists.
This involves making a copy of the list, which is inefficient
(at least compared to passing a numpy array) but these
return values are only really used in the testcases anyway.
Closes #891.