Remove duplicated checks.
Since the attribute is already checked for existence before calling
this function, there is almost no reason to check again here. The
only subtlety is that for derivatives this check will catch if the
derivative is infinite - but will erroneously report that the attribute
does not exist (also using the wrong attribute name). An infinite
derivative should not in any case prevent a new value from being set.