Move IMP_UNIQUE_PTR to more appropriate location.
It didn't work in compiler_macros.h because that header
is almost unique in that it does *not* include kernel_config.h,
which includes any #defines generated by cmake configure tests.
Thus the "does the compiler support unique_ptr" check always
evaluated to false in that header. Move to utility_macros.h
instead, which does pull in the neceesary #defines.