Mark decomposed restraints as used.
If a restraint decomposition results in a subrestraint that is
different from the original restraint but acts on the same set
of inputs, then it was not previously marked as used. This
resulted in a lot of "object was never used" warnings every time
an RMF frame was written. Should now be fixed.