Allow taking symmetry transform from a RigidBody.
Provide a new TransformationSymmetry constructor that takes
a RigidBody. The transformation used by the Modifier is then
taken from the reference frame of that RigidBody (and changes
if the reference frame does). The idea is to optimize the rigid body
transformation to infer the symmetry from input restraints.

Also provide a test function for the original TransformationSymmetry,
as this was previously missing.

Provided by Riccardo and Max, with minor cleanups by me.