Work around broken tagging of libimp_mpi.so DSO.
In Fedora 23 and newer, MPI DSOs are supposed to be tagged with
the MPI implementation. This doesn't seem to work correctly for
IMP (the Provides is tagged but the Requires is not, causing the
package to not install correctly). Work around this for now by
excluding libimp_mpi.so from requires *and* provides.