Don't fail if some bonds don't have CHARMMAtom.
Autogeneration of angles and dihedrals previously assumed
that the list of bonds given was generated by the CHARMM
forcefield code, such that every atom involved in a bond
had a CHARMM type (CHARMMAtom decorator), and would fail
if this was not the case. Now angles and dihedrals
involving atoms that don't have a CHARMM type are treated
in the same way as those for which CHARMM parameters
aren't available.