Add helpers to assert raising InternalExceptions.
Add new methods to the TestCase class, to be used
in any IMP testcase:
 - self.assertRaisesInternalException(callable)
 - self.assertNotImplemented(callable)
These will assert that the given callable object (e.g.
function) fails due to an internal check or because it
has been marked as not implemented. The assertions will
be automatically skipped if IMP is running without
internal checks turned on.

Also update current testcases to use the new helpers.