Give more useful found/not found dependency messages.
Previously, only the name of the dependency file was used
in cmake log messages indicating whether it was found or not.
This could be confusing in some cases - e.g. "OpenCV21 not found"
means that precisely version 2.1 of OpenCV could not be found,
while "OpenCV22 not found" means that OpenCV 2.2 or later was not
available. By allowing dependencies to give a longer human-readable
description, this can be clarified.