Check spelling in Python docstrings.
When we build the documentation, check each Python docstring
for spelling mistakes. This requires the Python enchant library
(and will be skipped if the library isn't available). Spelling
exceptions are taken from the standards_exceptions file in the
module's test directory; a new doc_spelling_exceptions variable
is now understood in that file, which is a list of all spelling
exceptions just for docstrings (as opposed to spelling_exceptions
which now applies to both docstrings and function/class names).