Checking for representation in subtrees, fixes #937.
expand_search now reports if it expanded representations.
Selection::search now stores whether it found a rep node up the tree.
Since the search is depth-first, that info is passed to the leaves.
At the bottom, checks if non-standard representation, and if so checks
that it encountered a Representation node. If not, returns nothing.