Remove timing-sensitive assertions.
Don't do benchmarking within test cases; they can be run in
parallel and so there is no guarantee that any test case (or
part of a test) gets 100% CPU. Benchmarks are the proper place
for such timing-sensitive measurements.

In any case, there is certainly no guarantee that an
IncrementalScoringFunction is any faster than a regular
RestraintsScoringFunction.