Allow subclassing ISDRestraint in Python.
In order to properly subclass in Python, SWIG directors must be
turned on for that class (by using the IMP_SWIG_BASE_OBJECT
macro rather than the regular IMP_SWIG_OBJECT). Otherwise, when
we try to subclass in Python what actually happens is that the
most derived class with directors turned on is used, which in
this case is the base IMP::Restraint, and so the wrong virtual
methods are called.