Set an unknown version if not using git.
If we got the develop source code from a tarball/zip file
rather than using git, it won't have a VERSION file in the
source directory and we can't use 'git describe' to create
one in the binary directory. Since other parts of the build
system require a VERSION file, set an unknown version in
this case. Closes #939.