Only use iterator constructor with real iterators.
The Vector(It, It) constructor is intended to be used
to construct a Vector from a pair of iterators, but since
it's templated, it'll match *any* usage where the types
of the two arguments are the same. Thus, Floats(1, 0)
will match (and then fail to compile with MSVC) rather
than being converted to Floats(1, 0.). Prevent the compiler
from using that constructor for any integral (non-iterator)
type. This means that we don't need to disambiguate
Floats(1, 0) any more.