Don't turn off missing-prototypes gcc warning.
Modern SWIG doesn't seem to generate wrappers
with the missing prototypes that gcc was complaining
about in the past. On the other hand, gcc6 warns when
we try to turn off this warning, because the wrappers
are mixed C++ and C code, and -Wmissing-prototypes
makes no sense for C++ code (since C++ is more strict
about prototypes).