Add a Refiner to get all leaves from a rigid body.
MultiFit's WeightedExcludedVolumeRestraint (which acts
between a number of rigid bodies) is written under the
assumption that each rigid body is also a Hierarchy, so
that it can use a LeavesRefiner on the rigid body to get
all leaves. That won't work if we use create_rigid_body()
instead of setup_as_rigid_body() since now the rigid body
is a separate particle from the hierarchy. Provide a new
RigidLeavesRefiner which returns all rigid members in a
given rigid body which are also hierarchy leaves. This can
be used in place of the LeavesRefiner.