Use context managers not RAII for temp directories.
Replace the IMP.test.TempDir and IMP.test.RunInTempDir
RAII objects with context managers (Python 'with'
statement). These should be more robust since the cleanup
code is triggered at the end of the 'with' block, not
when the object is destroyed. In regular CPython objects
can sometimes stick around for longer than expected
(if something has a reference to them) and in other Python
variants such as PyPy and Jython that use GC rather than
refcounting this 'extended lifetime' is almost guaranteed.