Add typemap workaround for older SWIG.
Older versions of SWIG don't realize that
::IMP::ParticleIndex (e.g. as emitted by the
IMP_DECORATOR_SETUP* macros) is the same as
IMP::ParticleIndex. This causes Python code that
tries to set up a decorator using a Particle or
Decorator (rather than a ParticleIndex) to fail,
e.g. in 6166d22.

Work around this by simply copying the ParticleIndex
typemaps.