Don't leak Model* in SAXS utility functions.
These functions create a Model but store it in a raw
pointer, which leaks. Since these functions create
Particles, the Model needs to live after the functions
exit, so a smart pointer can't be used internally. Fix this
by requiring the *caller* to provide a Model*. Update
FoXS and rg (users of these functions) to create a
Model and store it in a refcounted pointer. This should
also fix warnings from foxs that objects are not being
properly cleaned up on exit.