Move Model::get_particle() into header.
Profiling shows that this method gets called a lot.
Move the definition of this method (and its callee,
get_has_particle()) from the .cpp file to the .h
file to allow the compiler to optimize it a little.
Profiling shows that this reduces the time spent
calling this function by 2/3rds.