Make DirectionalDOF an IMP::Object.
Previously this class was a Value. Values can only be passed
by value or by const reference - it's not safe to pass a pointer
to them since they could disappear. Since DirectionalDOF stores
a fair amount of data, and *is* passed by pointer (e.g. to the
PathLocalPlanner constructor) it's more appropriate to make it
an Object.