Help SWIG with identical classes in two modules.
When a class with the same name is present in more than
one module (e.g. in this case, PCAFitRestraint in both
IMP::em and IMP::em2d) SWIG's %template function gets
confused and only generates code for one of the classes.
This causes IMP.em2d.PCAFitRestraint.get_from() to
not work properly. Work around this problem for now by
allowing IMP_SWIG_OBJECT to take an optional 4th argument,
which is a unique id that can be used to distinguish the
two classes for SWIG's sake. (Ideally SWIG would be smart
enough to handle this automatically, or at least to use
the existing provided Namespace argument, but in the latter
case it gets confused by "::".) Closes #985.