Exclude libmpi.so from requires.
RPM is supposed to tag any libmpi.so requires with the MPI library
(mpich in this case) but seems to do it inconsistently, so that the
final package ends up dependent on both
libmpi.so.12()(64bit) and libmpi.so.12()(64bit)(mpich-x86_64).
Since the mpich RPM only provides the latter, this makes the IMP-mpich
package uninstallable. Work around this by excluding the libmpi.so
requirement entirely. The package will still pull in the mpich RPM
since it has an explicit dependency.