Require system Eigen.
Add Eigen as a dependency rather than bundling a copy.
This allows us to easily use the latest version of Eigen,
getting any upstream fixes, without having to hack the
headers to put everything in a unique namespace and to
suppress compiler warnings. Additionally, it allows us
to use certain CGAL features which rely on system Eigen
being present. Relates #904.