Store provenance paths internally as absolute paths.
Since the current working directory may change during
an IMP run, it's more reliable to store any paths as
absolute paths. Note however that when written into
an RMF file, these will be converted into paths relative
to that of the RMF file, so that they don't break if the
directory containing the RMF is archived elsewhere.