Add the concept of an mmCIF component.
A component is similar to an IMP Chain, in that it
has a defined sequence and a unique ID (chain ID/asym_id),
but multiple Chains can map to one component (they can
model the same component in different states or even different
IMP Models) and a component can also represent something that
was not modeled (and so has no Chain). Since components must
have unique asym_ids, reassign Chain IDs if they are not unique.