Dump coordinates of starting models.
The mmCIF file contains the coordinates of the models used to
start the optimization, in PDB-like format. In the case where
the sequence differs between the input PDB file and the model
used by IMP (e.g. MSE mapped to MET) these mutations can also
be recorded in a suitable table.