Track all external modeling files.
mmCIF contains tables to track information that is stored in
external files - this can include Python scripts used, input
data files that aren't in official databases, and generated
files such as localization densities. These files can either be
on the local disk, or stored in a repository such as Zenodo.
Add a mechanism to map such files to mmCIF tables.