Add metadata to generated GMM text files
Add comments to any GMM file created with create_gmm.py
so that we can track back to the original MRC or PDB
file, and also see the options that create_gmm.py was
run with. This aids with reproducibility and makes it
easier to automate deposition of final models.