Add restraint filename key:value info
Allow for restraints to report arbitrary filename
information indexed by key, and store these filenames
in RMF files. Filenames are handled similarly to strings,
except that they are expected to be absolute paths, and
when written into the RMF they are converted into paths
relative to that of the RMF file.