Allow setting the GMM particle filename
If the density GMM particles were read from a file,
we should tell the restraint where this file is, so that
it can then report it as part of the RestraintInfo and
store it in RMF files. This allows the information to be tracked
back to the GMM file (and from there to the original EM density
file, by reading the metadata in the GMM file), which is
useful for deposition.