Suppress gcc8 warning about auto-generated SWIG code
gcc 8 will warn about a possible overflow when copying
into a fixed-length buffer. However, this code is OK,
so the warning is spurious. The code is also auto-generated
by SWIG, so we're not going to 'fix' it anyway. Suppress
the warning to avoid cluttering the output and potentially
obscuring other issues.