Use new sklearn GMM class if necessary
The sklearn folks have replaced the sklearn.mixture.GMM
class with a new GaussianMixture class. Use the new
class if the old one isn't available. The new class
is not a simple drop in replacement for the old one,
but appears to yield similar results. (In particular,
it's not possible to force radii or weights any more,
only to set their initial values.) Closes #981.