Allow for custom per-module cmake rules
Each module can currently contain a Setup.cmake file
which is included into the cmake scripts for that module.
However, this is intended to check for dependencies and
do other general setup. It is not intended to create custom
build rules, since these will run even if the module
is disabled (and so may fail if dependencies were not found).
Allow for an additional Build.cmake file which is only run
if the module is successfully enabled.