Bundle a copy of argparse for Python 2.6
This will allow us to use argparse to parse command
line arguments, rather than optparse, throughout the code.
optparse is deprecated and argparse is a more capable
replacement (but is not present in Python 2.6).
The argparse module is pretty small and pure Python, so
it is relatively straightforward to bundle with IMP
(and in most cases won't be used anyway, since many Python
2.6 systems include a backported argparse already).
Relates #991.