Allow overriding module name in out of tree builds
Normally for out of tree builds, the cmake scripts assume
that the last component of the path is the module name.
Allow overriding this. For example, this is useful when
building IMP.pmi1 out of tree, where the module name
(pmi1) doesn't match the default directory name (pmi)
after a clone of the PMI repository.