Pack the SphereD class
Rearrange the members of the SphereD class so
that the compiler adds no padding (previously, at
least with g++, 8 bytes of padding were added to
the start of the struct). This should save memory
and allow us to access sphere data from Python as
a simple 4*N array (for 3D spheres).