Return separate x,y,z,r NumPy arrays for spheres
Rather than trying to map a vector of Sphere3D objects
to a 2D NumPy array, which is very sensitive to how the
object is packed by the compiler, return separate x, y,
z and r arrays. This allows us to set the stride for
each array correctly and automatically handle padding.