Remove trivial Image2D copy constructor
Since our custom copy constructor doesn't
appear to do anything special, use the
default one instead (which may be faster
and is probably less error-prone since it
is guaranteed to copy everything). This will
also provide us with a default operator=
which we rely on elsewhere. (In C++11 the
instantiation of the default operator= is
deprecated with a custom copy constructor
and triggers a warning in gcc 9.)