Use detected Python, not "python"
Use the version of Python that CMake detected
to run the module setup scripts, rather than
assuming "python". This ensures that we use a
copy of Python that actually exists (on modern
systems there may not be a 'python', instead
only 'python2' or 'python3') and that matches
that used by the rest of the IMP build.