Rewrite shebang of Python apps
Rewrite the shebang line of each Python app
in the build directory to use the full Python
executable path as determined by CMake, rather
than assuming the first 'python' in the path.
This should fix running these apps on Python 3
systems, or if the path changes between config
time and run time.