Add missing tools and CMake config to imp-dev
The CMake scripts and configuration previously were not
installed by any of the Debian packages; install them as
part of imp-dev.

Tools, build information and SWIG interface files which
were previously in the main imp package are now moved
to imp-dev, as they are only needed when building against
IMP. Closes #1025.