Propagate Python errors through cleanup funcs
If we're doing cleanup as a result of a previous Python exception,
don't return a value from wrapper functions
(otherwise we'll get a SystemError "returned a result with
an error set" in recent versions of Python 3). Instead, return NULL
which will propagate the existing exception back to Python.

A common instance of this is yielding a set of decorators.
The last object may still be in scope at the end of the function
(at which point StopIteration is raised) and so its delete_xxx
wrapper may be called with StopException pending.