Don't call putback at EOF
Replace get/putback with peek to look at the next character
in the mol2 file. The previous code would call putback even
if the previous get failed (e.g. at EOF). On Linux this causes
the next call to getline to return the last line again, causing
an infinite loop.