Crop a DensityMap using a particle set
Allows cropping of an em:: DensityMap (such as a read-in .mrc file) using a
set of particles as the mask.

Voxels farther than the inputted distance threshold to all the particle
centers are set to a value of zero.

Setting "inverse" to false results in any voxels within the distance
threshold to any particles set to a value of zero.