Copy numpy arrays directly to Ints/Floats
Provide custom SWIG typemaps when built with
numpy support, that copy data directly from
the numpy array into an IMP Ints or Floats
array, provided the array is of the correct
native type and is laid out flat in memory.
This should be faster than the default typemaps,
particularly for large arrays, which enumerate
every item in the sequence and convert them to
and from Python objects.