More test fixes for particle index numpy arrays
Now that ParticleIndexes is returned as a NumPy
array, we can no longer use assertEqual to compare
two ParticleIndexes objects (as NumPy will do an
elementwise comparison, not want we want). Provide
an IMP.test method to compare two arrays which -
when we build with NumPy - uses numpy.testing to
compare the arrays instead. Also modify any use
of 'if pair in ParticleIndexPairs' since the 'in'
operator works differently in NumPy; the rough
equivalent might be
(pair==ParticleIndexPairs).all(axis=1).any()