Add basic Triggers to Model
Add a simple mechanism to Model to track when
to clear and rebuild caches of derived model
properties. Any IMP module can create Triggers
in either C++ or Python and the Model tracks
the time (in number of calls to Model::update())
when each Trigger was last updated. Restraints
or ScoreStates can rebuild their caches if these
trigger update times don't match the times when
the caches were last built.