Cache SoapDoublets calculation
Speed up OrientedSoapPairScore by roughly 50%
by caching the determination of atom doublets.
This is expensive because it requires traversing
the molecular hierarchy. Make sure the cache
is cleared if the hierarchy was modified since
the last scoring function evaluation (by monitoring
the core::Hierarchy::get_changed_key() trigger).
Benchmarking now shows that the score is comparable
to the implementation in Modeller (perhaps 5% faster).