Clear cache if Model changes
Since PairScores are not bound to a single
Model, it is possible that our cache is for
a different Model than the one we're being
called for. Fix this by simply clearing and
rebuilding the cache if the Model changes.
This is likely a rare scenario, so is more
straightforward than keeping a separate cache
for each Model.