Pass moved-particle info to restraint evaluation
Provide an alternative code path for evaluating restraints
that also takes a list of particles that have changed
since the last evaluation (Restraint::evaluate_moved()
and ScoringFunction::evaluate_moved()). This is intended
to be called by optimizers that can guarantee only a subset
of particles have moved (e.g. MonteCarlo). The default
implementation is the same as regular evaluate() but
Restraint or ScoringFunction subclasses can override this
to use the moved-particle information to score more efficiently.