Force full score at start/end of Monte Carlo
We shouldn't try to calculate the score of only
moved particles at the very start of an MC run
because we don't know what particles have moved
since the last evaluation (we can only guarantee
that *during* the MC run). Start with a full score.
Also, if we want to return the best score at the
end, make that a full score too.