Allow fragments in get_next/previous_residue
Expand the behavior of get_next_residue() and
get_previous_residue() so that they no longer
require that the Residue's immediate parent is
a Chain, only that a Chain is somewhere above
in the hierarchy. This should allow the functions
to work if the residues are in an intervening
Fragment, for example. Closes #933.