Add get_dependent_particles() utility function
Add a function that, given a single input particle,
uses the model dependency graph to return a list of
all particles that move when the input particle does.
In many cases this will just be the input particle,
but for example a rigid body particle as input will
give all body members in the output.