Add cached method to get dependent particles
Add a Model method to get all particles that
move when a given particle does. This might be
something like all members of a rigid body moving
when the body itself moves. This is done by querying
the dependency graph (e.g. for intervening ScoreStates).
This is expensive so the result is cached.