Don't symlink Python 2 to 3 __init__.py
We used to replace all Python 2 code with symlinks
to the Python 3 code, since it is the same. However,
in Ubuntu 22.04 or later, each module's __init__.py
might be different, since it contains configuration
information (e.g. the Python 3 build uses numpy but
the Python 2 build no longer does). So now we leave
those copies as is.