Use std::random_shuffle if it is available
For now, only use std::shuffle if
std::random_shuffle isn't provided. This should
fix builds on pre-C++11 platforms (std::shuffle
is a C++11 feature) and also on Ubuntu 22 (which
fails because it has gcc11, which requires that
min() and max() of the RNG are constexpr, but this
wasn't done in Boost until 1.75 and Ubuntu 22 has 1.74).