Assume 'final' keyword is always available
Now that we require C++11, we can use the 'final'
keyword. Keep the IMP_FINAL macro for now for
backwards compatibility. Provide an IMP_SWIG_FINAL
macro for the rare cases where we want to declare a
method final for C++ code but still allow SWIG to
override it for director classes.