Don't redistribute API sets
We shouldn't need to redistribute the API
sets (api-ms*.dll) or the universal C runtime,
as a) these should be installed on any up to
date Windows 7 system (or on Win10 by default)
and b) Microsoft recommends (or requires in some
docs) that they not be redistributed. At any rate,
most IMP users will have already installed Python 3,
which requires the same DLLs.