Allow save/restore of model dependencies
Add methods to Model to save and restore the
model dependency graph. These can be used to
bookend a code section that only temporarily
changes the graph, so that caches dependent
on the graph are not invalidated. Use these
methods when writing an RMF file, which can
create temporary decomposed restraints.