Build Mac Python extensions with 2-level namespace
Don't use -flat_namespace to build Python extensions,
as that causes problems when there are multiple libraries
available on Mac with the same name. For example, odd problems
occur with the conda package because IMP itself uses the
conda-provided libc++.1.dylib while the Python extensions use
the OS-provided libc++, and passing C++ objects between the
two does not (always) work. Use the default two-level namespace
instead so that we get the same library at runtime that we linked
against. Closes #1062.