Use correct score when RestraintSet is reset
When a particle is reset to its previous position
in move-aware scoring, we reset any restraints
involving that particle to their previous ("last-last")
scores. This did not work correctly when the restraint
was actually a RestraintSet, as the last-last score for
a RestraintSet was erroneously defined as the sum of
the last-last scores of its constituent restraints (this
is not always correct as some restraints may have been
reset and others rescored). Fix this by having the
RestraintSet keep its own last-last score, and invalidate
this (forcing a full recalculation) if restraints are
added to or removed from the set. Closes #1063.